requirements.txt
# Core dependencies for Knowledge-base Search Engine

# Web framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Document processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-magic==0.4.27

# Embeddings and ML
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1

# Vector storage
faiss-cpu==1.7.4
numpy==1.24.3

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# LLM integration
openai==1.3.5
tiktoken==0.5.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
chardet==5.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1